Fix handling of FQCNs in method parameters
Previously, MethodScanner did not detect a leading T_NS_SEPARATOR
character for a type name. By failing to detect it, it passes it to
ParameterScanner without it, who then dutifully resolves the classname
as if it was relative to the current namespace.

This is obviously incorrect behaviour, and it is fixed with this patch.
Tests have also been added and updated.