Dropping unused `laminas/laminas-eventmanager` dependency
Since `laminas/laminas-code:4.0.0`, we no longer have
code that relies on `laminas/laminas-eventmanager`.

The dependency is still dragged in though, and that
leads to indirectly installing `laminas/laminas-zendframework-bridge`,
which has a runtime performance impact on the entire autoloading
stack (see https://github.com/laminas/laminas-zendframework-bridge/issues/83 ).

Hereby we drop the dependency, which can safely be removed.