Deprecating `ParameterReflection#detectType()` and its callers
This API is old, bad, and not really useful, as it produces a string representation
of functions/methods that doesn't bring much to the table: in fact,
it is easier to inspect the AST of a symbol, than to look at its
string or array prototype.

These methods will be dropped in the next major release, as they
are prone to crashes due to union/intersection types on PHP 8.1
and newer, and it is not worth investing the time to fix them.

Signed-off-by: Marco Pivetta <ocramius@gmail.com>